@@ -1,11 +1,9 @@
 import { Injectable } from '@angular/core';
 import { User, Credentials, NewUser } from '../user';
-import { Observable } from 'rxjs/Observable';
+import { Observable , of } from 'rxjs';
 import { environment } from '../../environments/environment';
 import { HttpClient, HttpHeaders, HttpResponse, HttpErrorResponse } from '@angular/common/http';
 import { tap, catchError } from 'rxjs/operators';
-import { of } from 'rxjs/observable/of';
-import { ErrorObservable } from 'rxjs/observable/ErrorObservable';
 import { Router } from '@angular/router';
 
 const httpOptions = {
@@ -37,30 +35,30 @@ export class AuthService {
         this.isAuthenticated = false;
     }
 
-    public getCurrUser(): Observable<HttpResponse<User>> {
+    public getCurrUser(): Observable<User> {
         // return localStorage.getItem('auth-token') != null;
         const url = `${environment.apiUrl}/users/me`;
-        return this.http.get<HttpResponse<User>>(url, {observe: 'response'}).pipe(
+        return this.http.get<User>(url).pipe(
             tap((resp) => {
                 console.log(resp);
                 // this.setToken(resp.headers.get('Authorization'));
             }),
-            catchError(this.handleError)
+            catchError(this.handleError())
         );
     }
 
     // Returns the entire HTTP Response. JSON response
     // body is accessible via resp.body
-    public login(email: string, password: string): Observable<HttpResponse<User>> {
+    public login(email: string, password: string): Observable<User> {
         this.removeToken();
         const creds: Credentials = {email: email, password: password};
         const url = `${environment.apiUrl}/sessions`;
         console.log(url);
-        return this.http.post<HttpResponse<User>>(url, creds, {observe: 'response'}).pipe(
+        return this.http.post<User>(url, creds).pipe(
             tap((resp) => {
                 this.successfulSignIn(resp);
             }),
-            catchError(this.handleError)
+            catchError(this.handleError())
         );
     }
 
@@ -75,7 +73,7 @@ export class AuthService {
                 this.router.navigate(['/login']);
                 console.log('signed out!');
             }),
-            catchError(this.handleError)
+            catchError(this.handleError())
         );
     }
 
@@ -84,7 +82,7 @@ export class AuthService {
         password: string,
         passwordConf: string,
         firstName: string,
-        lastName: string): Observable<HttpResponse<User>> {
+        lastName: string): Observable<User> {
         const url = `${environment.apiUrl}/users`;
         const usr: NewUser = {
             email: email,
@@ -93,12 +91,12 @@ export class AuthService {
             firstName: firstName,
             lastName: lastName
         };
-        return this.http.post<HttpResponse<User>>(url, usr, {observe: 'response'})
+        return this.http.post<User>(url, usr)
         .pipe(
             tap((resp) => {
                 this.successfulSignIn(resp);
             }),
-            catchError(this.handleError)
+            catchError(this.handleError())
         );
     }
 
@@ -109,16 +107,20 @@ export class AuthService {
     }
 
     private log(message: string) {
+        // TODO: better logging, don't log to console
         console.log(message);
     }
 
-    private handleError(error: HttpErrorResponse) {
-        if (error.error instanceof ErrorEvent) {
-          // A client-side or network error occurred. Handle it accordingly.
-          console.error('An error occurred:', error.error.message);
-        }
-        this.removeToken();
-        // return an ErrorObservable with a user-facing error message
-        return new ErrorObservable(error);
+    private handleError<T> (operation = 'operation', result?: T) {
+        return (error: any): Observable<T> => {
+          // TODO: send the error to remote logging infrastructure
+          console.log('An error occurred:', error.error.message);
+
+          // TODO: better job of transforming error for user consumption
+          this.log(`${operation} failed: ${error.message}`);
+          this.removeToken();
+          // let the app keep running by returning an empty result
+          return of(result as T);
+        };
     }
 }
