@@ -2,47 +2,52 @@ import { Injectable } from '@angular/core';
 import { Release } from './release';
 import { environment } from '../environments/environment';
 import { Note, NewNote } from './note';
-import { HttpClient, HttpResponse, HttpErrorResponse} from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
+import { HttpClient} from '@angular/common/http';
+import { Observable, of} from 'rxjs';
 import { tap, catchError } from 'rxjs/operators';
-import { ErrorObservable } from 'rxjs/observable/ErrorObservable';
 
 @Injectable()
 export class NoteService {
 
     constructor(private http: HttpClient) {}
 
-    getNotes(release: Release): Observable<HttpResponse<Note[]>> {
+    getNotes(release: Release): Observable<Note[]> {
         const url = `${environment.apiUrl}/library/notes/releases/${release.id}`;
-        return this.http.get<HttpResponse<Note[]>>(url, {observe: 'response'}).pipe(
+        return this.http.get<Note[]>(url).pipe(
             tap((resp) => {
                 console.log(resp);
             }),
-            catchError(this.handleError)
+            catchError(this.handleError('getNotes', []))
         );
     }
 
-    postNote(note: string, release: Release): Observable<HttpResponse<Note>> {
+    postNote(note: string, release: Release): Observable<Note> {
         const url = `${environment.apiUrl}/library/notes/releases/${release.id}`;
         const noteObj: NewNote = {comment: note};
-        return this.http.post<HttpResponse<Note>>(url, noteObj, {observe: 'response'}).pipe(
+        return this.http.post<Note>(url, noteObj).pipe(
             tap((resp) => {
                 console.log(resp);
             }),
-            catchError(this.handleError)
+            catchError(this.handleError())
         );
 
     }
 
-
-
-    private handleError(error: HttpErrorResponse) {
-        if (error.error instanceof ErrorEvent) {
-          // A client-side or network error occurred. Handle it accordingly.
-          console.error('An error occurred:', error.error.message);
-        }
-        // return an ErrorObservable with a user-facing error message
-        return new ErrorObservable(error);
+    private log(message: string) {
+        // TODO: better logging, don't log to console
+        console.log(message);
     }
 
+    private handleError<T> (operation = 'operation', result?: T) {
+        return (error: any): Observable<T> => {
+          // TODO: send the error to remote logging infrastructure
+          console.log(error); // log to console instead
+
+          // TODO: better job of transforming error for user consumption
+          this.log(`${operation} failed: ${error.message}`);
+
+          // let the app keep running by returning an empty result
+          return of(result as T);
+        };
+    }
 }
